Refactoring the code
