+++ dgraph/cmd/migrate/run.go
@@ -21,7 +21,6 @@ import (
 	"os"
 
 	"github.com/dgraph-io/dgraph/x"
-	_ "github.com/go-sql-driver/mysql"
 	"github.com/spf13/cobra"
 )
